**Actual Behavior:**
Let's say there are 3 gateways. 300 is token value. 
20 rps means 2000 is bucket size.

|    | Gateway1  | Gateway2 | Gateway3| Decrement|
|  ---- | ---------- | ---------- |---------- |----------|
|1. request| 2000         | 2000         | 2000        |-300        |
|2. request| 1700         | 1700         | 1700        |-300         |
|3. request| 1400         | 1400         | 1400        |-300         |
|.| .         | .         | .        |.        |
|.| .         | .         | .        |.        |
|18. request| 200         | 200         | 200        |-300        |
|`Rate Limit exceeded`|

**Expected Behavior:**
Remaining bucket sizes are not used when they are single. They should be combined in a shared variable. 200+200+200 = 600/300 = 2 more requests should be done.

3 gateways. 20 rps.
20 should be saved to a shared variable. All gateway decrement it one by one.

As a shared variable, a Redis entry is used. I consider that latency increases. We may need a better way.

**Current Master:**
```
hey -c 20 -n 20 http://localhost:8080/drl/get

Summary:
  Total:	0.4010 secs
  Slowest:	0.4009 secs
  Fastest:	0.0107 secs
  Average:	0.2996 secs
  Requests/sec:	49.8765
  
  Total data:	84 bytes
  Size/request:	4 bytes

Response time histogram:
  0.011 [1]	|■■■
  0.050 [1]	|■■■
  0.089 [0]	|
  0.128 [0]	|
  0.167 [0]	|
  0.206 [6]	|■■■■■■■■■■■■■■■■■■■■
  0.245 [0]	|
  0.284 [0]	|
  0.323 [0]	|
  0.362 [0]	|
  0.401 [12]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■


Latency distribution:
  10% in 0.1943 secs
  25% in 0.1950 secs
  50% in 0.3999 secs
  75% in 0.4000 secs
  90% in 0.4006 secs
  95% in 0.4009 secs
  0% in 0.0000 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0040 secs, 0.0107 secs, 0.4009 secs
  DNS-lookup:	0.0011 secs, 0.0005 secs, 0.0016 secs
  req write:	0.0006 secs, 0.0001 secs, 0.0020 secs
  resp wait:	0.2946 secs, 0.0038 secs, 0.3967 secs
  resp read:	0.0002 secs, 0.0001 secs, 0.0005 secs

Status code distribution:
  [200]	18 responses
  [429]	2 responses
```

**With this PR:**
```
hey -c 20 -n 20 http://localhost:8080/drl/get

Summary:
  Total:	3.3064 secs
  Slowest:	3.3058 secs
  Fastest:	1.0609 secs
  Average:	1.9810 secs
  Requests/sec:	6.0489
  

Response time histogram:
  1.061 [1]	|■■■■
  1.285 [1]	|■■■■
  1.510 [3]	|■■■■■■■■■■■■■
  1.734 [9]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.959 [0]	|
  2.183 [0]	|
  2.408 [0]	|
  2.632 [1]	|■■■■
  2.857 [0]	|
  3.081 [0]	|
  3.306 [5]	|■■■■■■■■■■■■■■■■■■■■■■


Latency distribution:
  10% in 1.3443 secs
  25% in 1.5379 secs
  50% in 1.7050 secs
  75% in 3.0888 secs
  90% in 3.2673 secs
  95% in 3.3058 secs
  0% in 0.0000 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0053 secs, 1.0609 secs, 3.3058 secs
  DNS-lookup:	0.0015 secs, 0.0010 secs, 0.0018 secs
  req write:	0.0004 secs, 0.0000 secs, 0.0018 secs
  resp wait:	1.9750 secs, 1.0573 secs, 3.2995 secs
  resp read:	0.0002 secs, 0.0001 secs, 0.0004 secs

Status code distribution:
  [200]	20 responses
```

As it seems, this fixes the issue but total time changes `0.4010` secs to `3.3064` secs. It is a big change. So we may need another way for shared variable.

Fixes https://github.com/TykTechnologies/product/issues/230